requirements.txt
UnityPy==1.10.18
fsb5
wget
tqdm
Pillow
pyperclip
numpy
opencv-python